Move private useMentionSuggestions() hook into react-ui
