+++ packages/liveblocks-react-ui/src/primitives/Composer/index.tsx
@@ -13,7 +13,7 @@ import {
   size,
   useFloating,
 } from "@floating-ui/react-dom";
-import { type CommentBody, kInternal } from "@liveblocks/core";
+import type { CommentBody } from "@liveblocks/core";
 import { useRoomContextBundle } from "@liveblocks/react";
 import { Slot } from "@radix-ui/react-slot";
 import type {
@@ -60,6 +60,7 @@ import {
 
 import { useLiveblocksUIConfig } from "../../config";
 import { FLOATING_ELEMENT_COLLISION_PADDING } from "../../constants";
+import { useMentionSuggestions } from "../../shared";
 import { withAutoFormatting } from "../../slate/plugins/auto-formatting";
 import { withAutoLinks } from "../../slate/plugins/auto-links";
 import { withEmptyClearFormatting } from "../../slate/plugins/empty-clear-formatting";
@@ -624,10 +625,7 @@ const ComposerEditor = forwardRef<HTMLDivElement, ComposerEditorProps>(
     },
     forwardedRef
   ) => {
-    const {
-      [kInternal]: { useMentionSuggestions },
-      useSelf,
-    } = useRoomContextBundle();
+    const { useSelf } = useRoomContextBundle();
     const self = useSelf();
     const isDisabled = useMemo(
       () => disabled || !self?.canComment,
